pyproject.toml
[tool.poetry]
name = "teste-bringel"
version = "0.1.0"
description = ""
authors = ["weblerson <zonejr4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-oauth-toolkit = "^2.3.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
